aula-12/pyproject.toml
[project]
name = "aula-12"
version = "0.1.0"
description = ""
authors = [{ name = "Leonardo Souza", email = "leonardodisouza@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "faker (>=36.0.0,<37.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pandas (>=2.2.3,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "alive-progress (>=3.2.0,<4.0.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
